feat #1698 Moved aria.popup.PopupNavigationManager to aria.utils.DomNavigationManager
close #1698
